+++ Changelog.md
@@ -3,6 +3,7 @@
 ## Next
 
 * Introduces `Cursor::next_row` to fetch data without binding buffers first.
+* Removed `RowSetCursor::unbind` due to soundness issue, if fetching continues without any buffers bound.
 * Hides internal interface of `Cursor` trait behind `Cursor::stmt`.
 
 ## 0.16.0